requirements-docker.txt
boto3==1.9.38
certifi==2020.6.20
chardet==3.0.4
disclosure-extractor==0.0.47
eyed3
flask
ipython
PyPDF2==1.26.0
python-magic
idna==2.10
img2pdf
lxml==4.5.2
numpy==1.19.1
opencv-python==4.2.0.32
pandas==1.1.1
pdf2image==1.7.1
Pillow==8.0.1
pkginfo==1.5.0.1
PyPDF2==1.26.0
pytesseract==0.3.5
requests==2.25.0
six==1.15.0
urllib3==1.25.10
uwsgi
